+++ immutable-collection-encodings-test/src/test/java/com/hubspot/immutable/collection/encoding/test/ImmutableListEncodingTest.java
@@ -2,6 +2,9 @@ package com.hubspot.immutable.collection.encoding.test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.guava.GuavaModule;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Lists;
@@ -10,6 +13,10 @@ import java.util.List;
 import org.junit.Test;
 
 public class ImmutableListEncodingTest {
+
+  public static final String INPUT_JSON = "{}";
+  public static final ObjectMapper MAPPER = new ObjectMapper()
+    .registerModule(new GuavaModule());
 
   @Test
   public void itDoesNotCopyInputImmutableList() {
@@ -128,5 +135,11 @@ public class ImmutableListEncodingTest {
     TestList test = TestList.of(Collections.singleton("testing"));
 
     assertThat(test.getStrings()).containsExactly("testing");
+  }
+
+  @Test
+  public void itCanReadDefaultFromJsonMissingValue() throws JsonProcessingException {
+    TestListWithDefault test = MAPPER.readValue(INPUT_JSON, TestListWithDefault.class);
+    assertThat(test.getInts()).containsExactly(1);
   }
 }
